+++ src/main/java/com/livevox/challenge/app/agent/AgentDb.java
@@ -1,5 +1,6 @@
-package com.livevox.challenge.app;
+package com.livevox.challenge.app.agent;
 
+import com.livevox.challenge.app.agent.Agent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AgentDb extends JpaRepository<Agent, Long> {
