+++ src/main/java/com/livevox/challenge/app/call/CallDb.java
@@ -1,6 +1,6 @@
-package com.livevox.challenge.app;
+package com.livevox.challenge.app.call;
 
-import com.livevox.challenge.app.Call;
+import com.livevox.challenge.app.call.Call;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CallDb extends JpaRepository<Call, Long> {
